requirements_core.txt
# Image Processing  
Pillow>=10.0.0

# Data Science (optional UI helpers)
numpy>=1.24.0
pandas>=2.0.0

# Web Framework
streamlit>=1.35.0

# Utilities
requests>=2.31.0
python-dotenv>=1.0.0